now it's almost exactly like win11's navcontrol :))
