+++ WalletWasabi.Fluent/Controls/NavBarItem.axaml
@@ -56,7 +56,6 @@
     <Setter Property="VerticalAlignment" Value="Stretch" />
     <Setter Property="HorizontalAlignment" Value="Left" />
     <Setter Property="Width" Value="3" />
-    <Setter Property="Margin" Value="2 0 0 0"/>
     <Setter Property="Height" Value="NaN" />
   </Style>
